Centralize port connection checks and improve their reporting

+++ src/main/java/io/github/xfacthd/microredstone/client/screen/workbench/menu/ConstantPartNodeContextMenuProvider.java
@@ -102,7 +102,7 @@ public final class ConstantPartNodeContextMenuProvider extends PartNodeContextMe
         @Override
         public void saveQueryResult()
         {
-            node.setValue((short) editBox.getIntValue());
+            node.setValue(editBox.getIntValue());
         }
     }
 }
